refactor(core): extract consensus and task-execution modules (2/3)
Move the orchestration loop and the consensus subsystem out of orchestrator.ts
into two focused modules, with zero behavior change:

- task-execution.ts: the dependency-ordered execution loop (executeQueue),
  delegation team-info, checkpoint saves, and per-task prompt assembly
- consensus.ts: the proposer/judge verification loop (runConsensusCore) and the
  per-task verify hook (runTaskVerify)

orchestrator.ts drops from 3305 to 2213 lines. task-execution depends on
consensus (runTaskVerify); both build on the PR1 foundation modules. The module
graph stays acyclic.

Pure move: extracted bodies are byte-identical to the originals (only an export
prefix added) and the kept region is byte-identical. Verified with npm run lint,
1385 passing tests, zero diff under tests/, and a dist/index.d.ts identical to
the pre-split build.

Second of three sequential pure-move PRs; builds on the merged #397 foundation
modules.
